I fixed some import statements that were giving problems for the FeedView/FeedViewModel and for the DetailedFeedView/DetailedFeedViewModel. I also refactored my classes so that they are in the correct packages.
